Rewrite os module with pathlib and subrocess modules

+++ examples/run_notebooks.py
@@ -1,47 +1,47 @@
-import os
 import shutil
-
-cwd = os.getcwd()
+from pathlib import Path
+from subprocess import run
 
 clear = False
 pdf = False
 allow_errors = True
 
+exec_cmds = [
+    "jupyter",
+    "nbconvert",
+    "--execute",
+    "--ExecutePreprocessor.timeout=180000",
+]
+if allow_errors:
+    exec_cmds.append("--allow-errors")
+exec_cmds.append("--inplace")
+
 
 def run_nb(nb_file, nb_dir):
-    os.chdir(nb_dir)
-    worker_dirs = [
-        d for d in os.listdir(".") if os.path.isdir(d) and d.startswith("worker")
-    ]
-    for worker_dir in worker_dirs:
-        shutil.rmtree(worker_dir)
-    if allow_errors:
-        os.system(
-            "jupyter nbconvert --execute --ExecutePreprocessor.timeout=180000 --allow-errors --inplace {0}".format(
-                nb_file
-            )
-        )
-    else:
-        os.system(
-            "jupyter nbconvert --execute --ExecutePreprocessor.timeout=180000 --inplace {0}".format(
-                nb_file
-            )
-        )
+    for worker_dir in nb_dir.rglob("worker*"):
+        if worker_dir.isdir() and worker_dir.name.startswith("worker"):
+            shutil.rmtree(worker_dir)
+    run(exec_cmds + [nb_file], cwd=nb_dir, check=True)
     if pdf:
-        os.system("jupyter nbconvert --to pdf {0}".format(nb_file))
+        run(["jupyter", "nbconvert", "--to", "pdf", nb_file], cwd=nb_dir, check=True)
     if clear:
-        os.system(
-            "jupyter nbconvert --ClearOutputPreprocessor.enabled=True --ClearMetadataPreprocessor.enabled=True --allow-errors --inplace {0}".format(
-                nb_file
-            )
+        run(
+            [
+                "jupyter",
+                "nbconvert",
+                "--ClearOutputPreprocessor.enabled=True",
+                "--ClearMetadataPreprocessor.enabled=True",
+                "--inplace",
+                nb_file,
+            ],
+            cwd=nb_dir,
+            check=True,
         )
-    os.chdir(cwd)
     return
 
 
-nb_dir = "."
-nb_files = [f for f in os.listdir(nb_dir) if f.endswith(".ipynb")]
+nb_dir = Path(__file__).parent
+nb_files = sorted(nb_dir.rglob("*.ipynb"))
 assert len(nb_files) > 0
-nb_files.sort()
 for nb_file in nb_files:
     run_nb(nb_file, nb_dir)
